On branch ajustes2  Your branch is ahead of 'origin/ajustes2' by 1 commit.    (use "git push" to publish your local commits)
 Changes to be committed:
	modified:   .vscode/c_cpp_properties.json
	modified:   .vscode/launch.json
	modified:   .vscode/tasks.json
	modified:   Makefile
	new file:   bin/main_teste
	new file:   build/adm.o
	new file:   build/bibliotecario.o
	new file:   build/coluna_sql.o
	new file:   build/db_acervo.o
	new file:   build/db_administradores.o
	new file:   build/db_usuarios.o
	new file:   build/entidadebase.o
	new file:   build/livro.o
	new file:   build/usuario.o
	deleted:    main
	deleted:    main copy.cpp
	modified:   program/main.cpp
	new file:   program/main_teste.cpp
