+++ .vscode/tasks.json
@@ -2,7 +2,7 @@
     "tasks": [
         {
             "type": "cppbuild",
-            "label": "Tasks build PDS2-team3",
+            "label": "Tasks build biblioteca",
             "command": "/usr/bin/g++",
             "args": [
                 "-std=c++20",
@@ -15,7 +15,66 @@
                 "${workspaceFolder}/third_party",
                 "-lsqlite3",
                 "-o",
-                "${workspaceFolder}/main"
+                "${workspaceFolder}/bin/main_teste"
+            ],
+            "options": {
+                "cwd": "${fileDirname}"
+            },
+            "problemMatcher": [
+                "$gcc"
+            ],
+            "group": {
+                "kind": "build",
+                "isDefault": true
+            },
+            "detail": "Task generated by Debugger."
+        },
+        {
+            "type": "cppbuild",
+            "label": "Tasks build tester",
+            "command": "/usr/bin/g++",
+            "args": [
+                "-std=c++20",
+                "-g",
+                "${workspaceFolder}/src/*.cpp",
+                "${workspaceFolder}/tests/*.cpp",
+                "${workspaceFolder}/program/tester.cpp",
+                "-I",
+                "${workspaceFolder}/include",
+                "-I",
+                "${workspaceFolder}/third_party",
+                "-lsqlite3",
+                "-o",
+                "${workspaceFolder}/bin/test"
+            ],
+            "options": {
+                "cwd": "${fileDirname}"
+            },
+            "problemMatcher": [
+                "$gcc"
+            ],
+            "group": {
+                "kind": "build",
+                "isDefault": true
+            },
+            "detail": "Task generated by Debugger."
+        },
+        {
+            "type": "cppbuild",
+            "label": "Tasks build main teste",
+            "command": "/usr/bin/g++",
+            "args": [
+                "-std=c++20",
+                "-g",
+                "${workspaceFolder}/src/*.cpp",
+                "${workspaceFolder}/program/main_teste.cpp",
+                "-I",
+                "${workspaceFolder}/include",
+                "-I",
+                "${workspaceFolder}/third_party",
+                "-lsqlite3",
+                "-o",
+                "${workspaceFolder}/bin/main_teste"
             ],
             "options": {
                 "cwd": "${fileDirname}"
