+++ Makefile
@@ -15,6 +15,9 @@ TEST_OBJS = $(patsubst $(TESTDIR)%.cpp,$(OBJDIR)%.o,$(TEST_SRCS))
 TEST_MAIN = program/tester.cpp
 TEST_TARGET = bin/test
 
+MAIN_TEST_MAIN = program/main_teste.cpp
+MAIN_TEST_TARGET = bin/main_teste
+
 .PHONY: all clean test
 
 all: $(TARGET)
@@ -34,6 +37,12 @@ $(OBJDIR)/%.o: $(TESTDIR)/%.cpp
 test: $(TEST_TARGET)
 
 $(TEST_TARGET): $(OBJS) $(TEST_OBJS) $(TEST_MAIN)
+	@mkdir -p $(BINDIR)
+	$(CC) $(CFLAGS) -o $@ $^ -lsqlite3
+
+main_test: $(MAIN_TEST_TARGET)
+
+$(MAIN_TEST_TARGET): $(OBJS) $(MAIN_TEST_MAIN)
 	@mkdir -p $(BINDIR)
 	$(CC) $(CFLAGS) -o $@ $^ -lsqlite3
